refactor(#70): 회원가입 페이지 구조 리팩토링
