+++ libs/ui/src/lib/components/index.ts
@@ -6,7 +6,7 @@ export * from './DashboardCard';
 export * from './Input';
 export * from './ParticipationChart';
 export * from './SocialButton';
-export * from './signUp';
+export * from './signUp/input';
 export * from './ReceiveAiAssistant';
 export * from './LoadingOverlay';
 export * from './login';
