+++ libs/ui/src/lib/components/signUp/index.ts
@@ -1,6 +1,4 @@
-export * from './email';
-export * from './authcode';
-export * from './id';
-export * from './name';
-export * from './password';
-export * from './success';
+export * from './input';
+export * from './SignupHeader';
+export * from './SignupGuideText';
+export * from './SignupNavigation';
