+++ libs/ui/src/lib/components/signUp/input/index.ts
@@ -1,0 +1,6 @@
+export * from './email';
+export * from './authcode';
+export * from './id';
+export * from './name';
+export * from './password';
+export * from './success';
